pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "AI Agent with MCP and Skills abd more flexible support"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "mcp[cli]>=1.26.0",
    "openai>=2.26.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.1",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.30.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "core", "llm", "mcp_client", "models", "skills"]
